[API CHANGE] lok: Don't use 'bool' and 'uint64_t' in the stable API.
This is a f70e0ec6b3c61a7c7caa469949b0ac8016c89854 follow-up.

Reviewed-on: https://gerrit.libreoffice.org/42266
Tested-by: Michael Meeks <michael.meeks@collabora.com>

+++ desktop/source/lib/lokandroid.cxx
@@ -163,7 +163,7 @@ extern "C" SAL_JNI_EXPORT void JNICALL Java_org_libreoffice_kit_Office_setOption
 {
     LibreOfficeKit* pLibreOfficeKit = getHandle<LibreOfficeKit>(pEnv, aObject);
 
-    uint64_t pOptions = (uint64_t)options;
+    unsigned long long pOptions = (unsigned long long)options;
 
     pLibreOfficeKit->pClass->setOptionalFeatures(pLibreOfficeKit, pOptions);
 }
